Fix starcount, description and status update
reformat for linting and tests
commit before package structure change
