+++ gh_star_count/cli/count_stars_cli.py
@@ -2,7 +2,7 @@
 Command for counting number of stars in specified account
 """
 import typer
-from .count_stars import ask_prompt, count
+from .count_stars import main_starcount
 
 count_stars_cmd = typer.Typer(no_args_is_help=True)
 
@@ -12,8 +12,7 @@ def prompt_count():
     """
     prompt for account info and prints results from count
     """
-    name_search, orga_status = ask_prompt()
-    count(name_search, orga_status)
+    main_starcount()
 
 
 if __name__ == "__main__":
