+++ gh_star_count/cli/desc_update.py
@@ -1,7 +1,6 @@
 """
 Module allowing user description update on his github account
 """
-import os
 from github import Github
 
 
@@ -10,7 +9,10 @@ def get_user_token():
     Prompt the user to give login info
     """
     token = input(
-        "Github account access token (if you don't have one please visit https://github.com/settings/tokens) : "
+        (
+            "Github account access token (if you don't have one please visit:"
+            " https://github.com/settings/tokens) : "
+        )
     )
 
     return token
@@ -28,7 +30,7 @@ def update_user_bio(auth_user_obj):
     """
     update user bio with star, heart and tea emoji
     """
-    default = ":star: :gift_heart: :tea:"
+    default = ":star: :sunglasses: :tea:"
     print("-" * 80)
     print(f"default bio update : {default}")
     new_str = input("New bio for your account : ") or default
