+++ gh_star_count/cli/test_status_update.py
@@ -1,5 +1,3 @@
 """
 Unit tests for status update module
 """
-
-from .status_update import *
