+++ pyproject.toml
@@ -35,6 +35,7 @@ mkdocstrings = {extras = ["python"], version = "^0.19.0"}
 pytest-cov = "^4.0.0"
 mkdocs-coverage = "^0.2.5"
 git-changelog = "^0.6.0"
+dotenv = "^0.0.5"
 
 [build-system]
 requires = ["poetry-core"]
